pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Bryce McIntyre <brycemcintyre@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.2"
"discord.py" = {extras = ["voice"], version = "^1.6.0"}
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
